requirements-test.txt
pytest
pytest-homeassistant-custom-component>=0.12.8
pytest-asyncio
pytest-cov
boto3>=1.28.0 